requirements.txt
django==5.2.7
django-allauth==65.12.0
python-dotenv==1.1.1
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.3

